[fix] stutter when leaving FreeCam
I checked with #1759 and was not able to reproduce the issue.

This problem is exaggerated when we call the loadRenderers with OptiFine added.

We do not need to reload the entire rendering pipeline every time we leave FreeCam.

If we still encounter the issue, I would much rather block unloading chunks from the graphics pipeline and check to make sure that works with OptiFine.